requirements.txt
kafka-python==2.0.2
oss2==2.16.0
